Fix line artifacts.
The issue was in the paint method; it wasn't drawing enough grid lines.
The issue is that the paint_rect tends not to be exactly aligned to the
grid (it tends to be just a little bit larger). This means that the
width and height of the paint_rect are not good indicators of the number
of squares that need to be redrawn.